＊コードはファイルを分割する ＊ナビゲーションパターン設定する ＊Widgetとページコードはバックエンド処理から分けるの準備が始まる
