+++ lib/Common/routes.dart
@@ -1,0 +1,5 @@
+class Routes {
+  
+  static const String search = '/search';
+
+}
