+++ lib/app.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:gitsearch/Common/routes.dart';
+import 'package:gitsearch/Pages/search_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+
+    return MaterialApp(
+      title: 'GitSearch',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      onGenerateRoute: (RouteSettings settings){
+
+        if (kDebugMode) {
+          print('Route: ${settings.name}');
+        }
+
+        switch(settings.name){
+          case Routes.search:
+            return MaterialPageRoute<dynamic>(
+                builder: (_) => const SearchPage(),
+                settings: settings,
+              );
+          default:
+            return MaterialPageRoute<dynamic>(
+                builder: (_) => const SearchPage(),
+                settings: settings,
+              );
+        }
+
+      }
+      
+    );
+  }
+}
